Clean  component attribute class
